Added date code and decode. Also adapted add payroll and transaction

+++ src/models/payroll.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub struct Payroll {
     _id: i32,
-    pub date: i32,
+    pub date: i64,
     pub gross: f32,
     pub net: f32,
     pub ss: f32,
@@ -11,7 +11,7 @@ pub struct Payroll {
 }
 
 impl Payroll {
-    pub fn new(date: i32, gross: f32, net: f32, ss: f32, irpf: f32, company_id: i32, category_id: i32) -> Payroll {
+    pub fn new(date: i64, gross: f32, net: f32, ss: f32, irpf: f32, company_id: i32, category_id: i32) -> Payroll {
         Payroll {
             _id: 0,
             date,
